ict_resources_backend/requirements.txt
asgiref==3.2.10
appdirs==1.4.3
distlib==0.3.0
Django==3.0.5
django-jazzmin
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
filelock==3.0.12
importlib-metadata==1.6.0
Pillow==7.1.1
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.15
zipp==3.1.0

rest_social_auth
django-anymail[mailgun]
django_seed
django_q
django-data-seeder
social-auth-app-django
